+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### HEAD
 
+- Added a `isDataURI` validator
+  ([#521](https://github.com/chriso/validator.js/pull/521))
 - Fixed a bug with `isURL()` when protocol was missing and :// appeared in the query
   ([#518](https://github.com/chriso/validator.js/issues/518))